Move hydraulic machines into new folder
