+++ src/main/java/io/github/pylonmc/pylon/base/PylonBlocks.java
@@ -5,12 +5,12 @@ import io.github.pylonmc.pylon.base.fluid.pipe.FluidPipeMarker;
 import io.github.pylonmc.pylon.base.items.DimensionalBarrel;
 import io.github.pylonmc.pylon.base.items.ExplosiveTarget;
 import io.github.pylonmc.pylon.base.items.Immobilizer;
-import io.github.pylonmc.pylon.base.items.hydraulic.HydraulicGrindstoneTurner;
-import io.github.pylonmc.pylon.base.items.hydraulic.HydraulicHammerHead;
-import io.github.pylonmc.pylon.base.items.hydraulic.HydraulicMixingAttachment;
+import io.github.pylonmc.pylon.base.items.hydraulic.machines.HydraulicGrindstoneTurner;
+import io.github.pylonmc.pylon.base.items.hydraulic.machines.HydraulicHammerHead;
+import io.github.pylonmc.pylon.base.items.hydraulic.machines.HydraulicMixingAttachment;
 import io.github.pylonmc.pylon.base.items.Press;
 import io.github.pylonmc.pylon.base.items.fluid.*;
-import io.github.pylonmc.pylon.base.items.hydraulic.HydraulicPressPiston;
+import io.github.pylonmc.pylon.base.items.hydraulic.machines.HydraulicPressPiston;
 import io.github.pylonmc.pylon.base.items.multiblocks.*;
 import io.github.pylonmc.pylon.base.items.multiblocks.smelting.*;
 import io.github.pylonmc.pylon.base.items.tools.watering.Sprinkler;
