+++ src/main/java/io/github/pylonmc/pylon/base/PylonEntities.java
@@ -3,11 +3,11 @@ package io.github.pylonmc.pylon.base;
 import io.github.pylonmc.pylon.base.fluid.pipe.FluidPipeDisplay;
 import io.github.pylonmc.pylon.base.fluid.pipe.connection.FluidConnectionDisplay;
 import io.github.pylonmc.pylon.base.fluid.pipe.connection.FluidConnectionInteraction;
-import io.github.pylonmc.pylon.base.items.hydraulic.HydraulicHammerHead;
-import io.github.pylonmc.pylon.base.items.hydraulic.HydraulicMixingAttachment;
+import io.github.pylonmc.pylon.base.items.hydraulic.machines.HydraulicHammerHead;
+import io.github.pylonmc.pylon.base.items.hydraulic.machines.HydraulicMixingAttachment;
 import io.github.pylonmc.pylon.base.items.Press;
 import io.github.pylonmc.pylon.base.items.fluid.*;
-import io.github.pylonmc.pylon.base.items.hydraulic.HydraulicPressPiston;
+import io.github.pylonmc.pylon.base.items.hydraulic.machines.HydraulicPressPiston;
 import io.github.pylonmc.pylon.base.items.multiblocks.Grindstone;
 import io.github.pylonmc.pylon.base.items.multiblocks.Pedestal;
 import io.github.pylonmc.pylon.base.items.multiblocks.smelting.SmelteryController;
